GUARD-3717 2 Do a couple of cleanup TODOs
