+++ src/ShopifyAccess/ShopifyAccess.csproj
@@ -101,6 +101,7 @@
     <Compile Include="GraphQl\Queries\GetProductVariantsInventoryReportQuery.cs" />
     <Compile Include="GraphQl\Queries\QueryBuilder.cs" />
     <Compile Include="GraphQl\ReportType.cs" />
+    <Compile Include="GraphQl\Services\GraphQlPaginationService.cs" />
     <Compile Include="GraphQl\Services\IReportGenerator.cs" />
     <Compile Include="GraphQl\Services\ReportGenerator.cs" />
     <Compile Include="GraphQl\ShopifyGraphQlThrottler.cs" />
