+++ src/ShopifyAccessTests/ShopifyAccessTests/ShopifyAccessTests.csproj
@@ -110,7 +110,6 @@
     <Compile Include="Orders\OrderMapperTests.cs" />
     <Compile Include="Orders\OrdersListTests.cs" />
     <Compile Include="Orders\OrdersTests.cs" />
-    <Compile Include="Products\Models\ShopifyProductVariantTests.cs" />
     <Compile Include="Products\ProductVariantTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\WebRequestServicesTests.cs" />
